extracted backup logic into seperate service
