+++ src/Altinn.Platform/Altinn.Platform.Storage/DataCleanup/Services/BlobService.cs
@@ -73,92 +73,7 @@ namespace Altinn.Platform.Storage.DataCleanup.Services
 
             return true;
         }
-
-        /// <inheritdoc/>
-        public async Task<bool> DeleteInstanceBackup(string instanceOwnerPartyId, string instanceGuid)
-        {
-            BlobContainerClient container = await CreateBackupBlobClient();
-
-            try
-            {
-                await foreach (BlobItem item in container.GetBlobsAsync(BlobTraits.None, BlobStates.None, $"instances/{instanceOwnerPartyId}/{instanceGuid}", CancellationToken.None))
-                {
-                    container.DeleteBlobIfExists(item.Name, DeleteSnapshotsOption.IncludeSnapshots);
-                }
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "BlobService // DeleteInstanceBackup // Instance: {InstanceOwnerPartyId}/{InstanceGuid} // Exeption: {Exception}", instanceOwnerPartyId, instanceGuid, e);
-                return false;
-            }
-
-            return true;
-        }
-
-        /// <inheritdoc/>
-        public async Task<bool> DeleteInstanceEventsBackup(string instanceOwnerPartyId, string instanceGuid)
-        {
-            BlobContainerClient container = await CreateBackupBlobClient();
-
-            try
-            {
-                await foreach (BlobItem item in container.GetBlobsAsync(BlobTraits.None, BlobStates.None, $"instanceEvents/{instanceOwnerPartyId}/{instanceGuid}", CancellationToken.None))
-                {
-                    container.DeleteBlobIfExists(item.Name, DeleteSnapshotsOption.IncludeSnapshots);
-                }
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "BlobService // DeleteInstanceEventsBackup // Instance: {InstanceOwnerPartyId}/{InstanceGuid} // Exeption: {Exception}", instanceOwnerPartyId, instanceGuid, e);
-                return false;
-            }
-
-            return true;
-        }
-
-        /// <inheritdoc/>
-        public async Task<bool> DeleteDataBackup(string instanceGuid)
-        {
-            BlobContainerClient container = await CreateBackupBlobClient();
-
-            try
-            {
-                await foreach (BlobItem item in container.GetBlobsAsync(BlobTraits.None, BlobStates.None, $"dataElements/{instanceGuid}", CancellationToken.None))
-                {
-                    container.DeleteBlobIfExists(item.Name, DeleteSnapshotsOption.IncludeSnapshots);
-                }
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "BlobService // DeleteInstanceEventsBackup // Instance: {InstanceGuid} // Exeption: {Exception}", instanceGuid, e);
-                return false;
-            }
-
-            return true;
-        }
-
-        private async Task<BlobContainerClient> CreateBackupBlobClient()
-        {
-            // local emulator
-            if (_accountName.Equals("devstoreaccount1"))
-            {
-                StorageSharedKeyCredential credentials = new StorageSharedKeyCredential(_accountName, _accountKey);
-                BlobServiceClient client = new BlobServiceClient(new Uri(_blobEndpoint), credentials);
-                return client.GetBlobContainerClient("backup");
-            }
-
-            string backupAccountKey = await _keyVaultService.GetSecretAsync(
-            string.Format(_vaultUri, _environment),
-            "AzureStorageConfiguration--BackupAccountKey");
-
-            StorageSharedKeyCredential storageCredentials = new StorageSharedKeyCredential(_backupAccountName, backupAccountKey);
-            Uri storageUrl = new Uri(string.Format(_backupAccountEndpoint, _environment));
-
-            BlobServiceClient commonBlobClient = new BlobServiceClient(storageUrl, storageCredentials);
-            BlobContainerClient blobContainerClient = commonBlobClient.GetBlobContainerClient("backup");
-            return blobContainerClient;
-        }
-
+       
         private async Task<BlobContainerClient> CreateBlobClient(string org)
         {
             if (!_accountName.Equals("devstoreaccount1"))
