+++ src/Altinn.Platform/Altinn.Platform.Storage/DataCleanup/host.json
@@ -12,6 +12,7 @@
       "Function": "Error",
       "Altinn.Platform.Storage.DataCleanup.Services.CosmosService": "Error",
       "Altinn.Platform.Storage.DataCleanup.Services.BlobService": "Error",
+      "Altinn.Platform.Storage.DataCleanup.Services.BackupBlobService": "Error",
       "Altinn.Platform.Storage.DataCleanup.Services.SasTokenProvider": "Error",
       "Altinn.Platform.Storage.DataCleanup.Services.KeyVaultService": "Error"
     }
